Moved node logic to /node specific files
