+++ src/pprocess.c
@@ -16,14 +16,6 @@
 
 enum role global_role = ROLE_INGRESS;
 
-uint8_t k_pot_in[SID_NO][HMAC_MAX_LENGTH] = {
-    // egress key (16 bytes), pad with zeros if HMAC_MAX_LENGTH > 16
-    {0xa1, 0xb2, 0xc3, 0xd4, 0xe5, 0xf6, 0x07, 0x18, 0x29, 0x30, 0x4b, 0x6c, 0x7d, 0x8e, 0x9f, 0xa0,
-     0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0},
-    // transit key (16 bytes), pad with zeros if needed
-    {0xd8, 0xf9, 0xcd, 0xe1, 0xab, 0x34, 0x5c, 0xd0, 0xef, 0x67, 0x89, 0xab, 0x12, 0xcd, 0xef, 0x34,
-     0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0,    0}};
-
 void remove_headers(struct rte_mbuf *pkt) {
   struct rte_ether_hdr *eth_hdr_6 = rte_pktmbuf_mtod(pkt, struct rte_ether_hdr *);
   struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr_6 + 1);
@@ -296,586 +288,6 @@ void add_custom_header_only(struct rte_mbuf *pkt) {
   RTE_LOG(INFO, USER1, "Custom headers added to packet\n");
 }
 
-// Add this function to handle return traffic at the ingress node
-static inline void process_ingress_return_packet(struct rte_mbuf *mbuf) {
-  struct rte_ether_hdr *eth_hdr = rte_pktmbuf_mtod(mbuf, struct rte_ether_hdr *);
-  struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-
-  printf("**** [INGRESS-RETURN] DEBUG: Function called with packet length %u ****\n", rte_pktmbuf_pkt_len(mbuf));
-  printf("[INGRESS-RETURN] MAC src: %02X:%02X:%02X:%02X:%02X:%02X\n", 
-         eth_hdr->src_addr.addr_bytes[0], eth_hdr->src_addr.addr_bytes[1],
-         eth_hdr->src_addr.addr_bytes[2], eth_hdr->src_addr.addr_bytes[3],
-         eth_hdr->src_addr.addr_bytes[4], eth_hdr->src_addr.addr_bytes[5]);
-  printf("[INGRESS-RETURN] MAC dst: %02X:%02X:%02X:%02X:%02X:%02X\n",
-         eth_hdr->dst_addr.addr_bytes[0], eth_hdr->dst_addr.addr_bytes[1],
-         eth_hdr->dst_addr.addr_bytes[2], eth_hdr->dst_addr.addr_bytes[3],
-         eth_hdr->dst_addr.addr_bytes[4], eth_hdr->dst_addr.addr_bytes[5]);
-  
-  char src_ip[INET6_ADDRSTRLEN], dst_ip[INET6_ADDRSTRLEN];
-  inet_ntop(AF_INET6, &ipv6_hdr->src_addr, src_ip, INET6_ADDRSTRLEN);
-  inet_ntop(AF_INET6, &ipv6_hdr->dst_addr, dst_ip, INET6_ADDRSTRLEN);
-  
-  printf("[INGRESS-RETURN] IPv6 src: %s\n", src_ip);
-  printf("[INGRESS-RETURN] IPv6 dst: %s\n", dst_ip);
-  printf("[INGRESS-RETURN] Next Header: %u\n", ipv6_hdr->proto);
-  
-  // Forward directly to client
-  struct in6_addr client_addr;
-  inet_pton(AF_INET6, "2a05:d014:dc7:12ff:f611:cc26:cf0d:5c92", &client_addr);
-  struct rte_ether_addr *client_mac = lookup_mac_for_ipv6(&client_addr);
-
-  if (client_mac) {
-    printf("[INGRESS-RETURN] Found client MAC: %02X:%02X:%02X:%02X:%02X:%02X\n",
-           client_mac->addr_bytes[0], client_mac->addr_bytes[1],
-           client_mac->addr_bytes[2], client_mac->addr_bytes[3],
-           client_mac->addr_bytes[4], client_mac->addr_bytes[5]);
-    send_packet_to(*client_mac, mbuf, 0);
-    printf("[INGRESS-RETURN] Forwarded server response to client\n");
-  } else {
-    printf("[INGRESS-RETURN] No MAC mapping for client! Dropping packet\n");
-    rte_pktmbuf_free(mbuf);
-  }
-}
-
-// Helper: process a single packet for ingress
-static inline void process_ingress_packet(struct rte_mbuf *mbuf, uint16_t rx_port_id) {
-  printf("[INGRESS] Packet hex dump BEFORE processing (first 64 bytes):\n");
-  size_t dump_len = rte_pktmbuf_pkt_len(mbuf);
-  if (dump_len > 64) dump_len = 64;
-  hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len);
-  printf("[INGRESS] Packet length: %u\n", rte_pktmbuf_pkt_len(mbuf));
-  printf("[INGRESS] mbuf nb_segs: %u\n", mbuf->nb_segs);
-  printf("[INGRESS] tailroom: %u, data_len: %u\n", rte_pktmbuf_tailroom(mbuf), mbuf->data_len);
-
-  struct rte_ether_hdr *eth_hdr = rte_pktmbuf_mtod(mbuf, struct rte_ether_hdr *);
-  uint16_t ether_type = rte_be_to_cpu_16(eth_hdr->ether_type);
-
-  // 1. Skip non-IPv6
-  if (ether_type != RTE_ETHER_TYPE_IPV6) {
-    printf("Ingress: Not an IPv6 packet, skipping and freeing mbuf\n");
-    rte_pktmbuf_free(mbuf);
-    return;
-  }
-
-  // 2. Skip non-unicast (multicast/broadcast) destination MAC
-  if ((eth_hdr->dst_addr.addr_bytes[0] & 0x01) != 0) {
-    printf("Ingress: Multicast/Broadcast MAC, skipping and freeing mbuf\n");
-    // Print more info for debugging
-    printf("  MAC: %02X:%02X:%02X:%02X:%02X:%02X\n", eth_hdr->dst_addr.addr_bytes[0],
-           eth_hdr->dst_addr.addr_bytes[1], eth_hdr->dst_addr.addr_bytes[2], eth_hdr->dst_addr.addr_bytes[3],
-           eth_hdr->dst_addr.addr_bytes[4], eth_hdr->dst_addr.addr_bytes[5]);
-    struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-    printf("  IPv6 src: ");
-    print_ipv6_address((struct in6_addr *)&ipv6_hdr->src_addr, "");
-    printf("  IPv6 dst: ");
-    print_ipv6_address((struct in6_addr *)&ipv6_hdr->dst_addr, "");
-    printf("  EtherType: 0x%04x\n", ether_type);
-    printf("  Next header: %u\n", ipv6_hdr->proto);
-    if (ipv6_hdr->proto == IPPROTO_UDP) {
-      struct rte_udp_hdr *udp_hdr = (struct rte_udp_hdr *)(ipv6_hdr + 1);
-      printf("  UDP src port: %u, dst port: %u\n", ntohs(udp_hdr->src_port), ntohs(udp_hdr->dst_port));
-    } else if (ipv6_hdr->proto == IPPROTO_TCP) {
-      struct rte_tcp_hdr *tcp_hdr = (struct rte_tcp_hdr *)(ipv6_hdr + 1);
-      printf("  TCP src port: %u, dst port: %u\n", ntohs(tcp_hdr->src_port), ntohs(tcp_hdr->dst_port));
-    }
-    rte_pktmbuf_free(mbuf);
-    return;
-  }
-
-  switch (ether_type) {
-    case RTE_ETHER_TYPE_IPV6:
-      switch (operation_bypass_bit) {
-        case 0:
-          printf("Processing IPv6 packet with operation_bypass_bit = 0\n");
-          add_custom_header(mbuf);
-          // Print after header addition
-          printf("[INGRESS] After header addition: Packet length: %u\n", rte_pktmbuf_pkt_len(mbuf));
-          printf("[INGRESS] After header addition: mbuf nb_segs: %u\n", mbuf->nb_segs);
-          printf("[INGRESS] After header addition: tailroom: %u, data_len: %u\n", rte_pktmbuf_tailroom(mbuf),
-                 mbuf->data_len);
-
-          // Realign pointers after header addition
-          printf("Realigning pointers after adding custom headers\n");
-          struct rte_ether_hdr *eth_hdr6 = rte_pktmbuf_mtod(mbuf, struct rte_ether_hdr *);
-          struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr6 + 1);
-          struct ipv6_srh *srh = (struct ipv6_srh *)(ipv6_hdr + 1);
-          struct hmac_tlv *hmac = (struct hmac_tlv *)(srh + 1);
-          struct pot_tlv *pot = (struct pot_tlv *)(hmac + 1);
-
-          // Print offsets and hex dumps for debugging
-          print_offset_and_hex("[INGRESS] SRH", eth_hdr6, srh, sizeof(struct ipv6_srh));
-          print_offset_and_hex("[INGRESS] HMAC TLV", eth_hdr6, hmac, sizeof(struct hmac_tlv));
-          print_offset_and_hex("[INGRESS] POT TLV", eth_hdr6, pot, sizeof(struct pot_tlv));
-
-          // Extract destination IPv6 address as string
-          printf("Extracting destination IPv6 address\n");
-          char dst_ip_str[INET6_ADDRSTRLEN];
-          if (inet_ntop(AF_INET6, &ipv6_hdr->dst_addr, dst_ip_str, sizeof(dst_ip_str)) == NULL) {
-            perror("inet_ntop failed");
-            break;
-          }
-
-          // Prepare k_pot_in array using the loaded key for all SIDs
-          // printf("Preparing k_pot_in array for all SIDs\n");
-          // uint8_t k_pot_in[SID_NO][HMAC_MAX_LENGTH] = {0};
-          // for (int sid = 0; sid < SID_NO; sid++) {
-          //   if (read_encryption_key("keys.txt", dst_ip_str, k_pot_in[sid], HMAC_MAX_LENGTH) != 0) {
-          //     printf("Failed to read key for %s\n", dst_ip_str);
-          //     break;
-          //   }
-          // }
-
-          printf("Debug: k_pot_in array contents:\n");
-          for (int i = 0; i < SID_NO; i++) {
-            printf("SID %d: ", i);
-            for (int j = 0; j < HMAC_MAX_LENGTH; j++) {
-              printf("%02x", k_pot_in[i][j]);
-            }
-            printf("\n");
-          }
-
-          // Prepare HMAC key (use same as encryption key for demo)
-          // printf("Preparing HMAC key for HMAC computation\n");
-          // uint8_t k_hmac_ie[HMAC_MAX_LENGTH] = {0};
-          // if (read_encryption_key("keys.txt", dst_ip_str, k_hmac_ie, HMAC_MAX_LENGTH) != 0) {
-          //   printf("Failed to read HMAC key for %s\n", dst_ip_str);
-          //   break;
-          // }
-          uint8_t *k_hmac_ie = k_pot_in[0];  // egress key
-          size_t key_len = HMAC_MAX_LENGTH;
-
-          // Compute HMAC
-          printf("Computing HMAC for the packet\n");
-          // Log HMAC input values for debugging
-          struct in6_addr ingress_addr;
-          inet_pton(AF_INET6, "2a05:d014:dc7:12ff:f611:cc26:cf0d:5c92", &ingress_addr);
-          printf("[INGRESS] HMAC input values (FORCED ingress IP):\n");
-          printf("  src_addr: 2a05:d014:dc7:12ff:f611:cc26:cf0d:5c92\n");
-          printf("  srh->last_entry: %u\n", srh->last_entry);
-          printf("  srh->flags: %u\n", srh->flags);
-          printf("  hmac->hmac_key_id: %u\n", rte_be_to_cpu_32(hmac->hmac_key_id));
-          printf("  srh->segments: ");
-          for (int i = 0; i <= srh->last_entry; i++) {
-            char seg_str[INET6_ADDRSTRLEN];
-            if (inet_ntop(AF_INET6, &srh->segments[i], seg_str, sizeof(seg_str)))
-              printf("%s ", seg_str);
-            else
-              perror("inet_ntop segment");
-          }
-          printf("\n");
-          // Dump the first 128 bytes (or the whole packet if smaller) after all headers and before send
-          size_t dump_len = rte_pktmbuf_pkt_len(mbuf);
-          if (dump_len > 128) dump_len = 128;
-          printf("[INGRESS] Packet hex dump before send (first %zu bytes):\n", dump_len);
-          hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len);
-          uint8_t hmac_out[HMAC_MAX_LENGTH];
-          if (calculate_hmac((uint8_t *)&ingress_addr, srh, hmac, k_hmac_ie, key_len, hmac_out) == 0) {
-            printf("HMAC Computation Successful\nHMAC: ");
-            for (int i = 0; i < HMAC_MAX_LENGTH; i++) printf("%02x", hmac_out[i]);
-            printf("\n");
-            rte_memcpy(hmac->hmac_value, hmac_out, HMAC_MAX_LENGTH);
-            printf("HMAC value inserted to srh_hmac header\n");
-
-            // Dump the first 128 bytes (or the whole packet if smaller) after HMAC is written
-            size_t dump_len_after_hmac = rte_pktmbuf_pkt_len(mbuf);
-            if (dump_len_after_hmac > 128) dump_len_after_hmac = 128;
-            printf("[INGRESS] Packet hex dump AFTER HMAC write (first %zu bytes):\n", dump_len_after_hmac);
-            hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len_after_hmac);
-            printf("HMAC value inserted to srh_hmac header\n");
-          } else {
-            printf("HMAC Computation Failed\n");
-            break;
-          }
-
-          // Nonce Generation and Encryption (PVF Computation)
-          printf("Generating nonce and encrypting PVF\n");
-          uint8_t nonce[NONCE_LENGTH];
-          if (generate_nonce(nonce) != 0) {
-            printf("Nonce generation failed, returning\n");
-            break;
-          }
-          encrypt_pvf(k_pot_in, nonce, hmac_out);
-          printf("Encrypted PVF before writing to the header: ");
-          for (int i = 0; i < HMAC_MAX_LENGTH; i++) printf("%02x", hmac_out[i]);
-          printf("\n");
-          rte_memcpy(pot->encrypted_hmac, hmac_out, HMAC_MAX_LENGTH);
-          rte_memcpy(pot->nonce, nonce, NONCE_LENGTH);
-          printf("Encrypted PVF and nonce values inserted to pot header\n");
-
-          // Forward the packet using the `send_packet_to` function
-          printf("Forwarding packet to next hop\n");
-
-          if (srh->segments_left == 0) {
-            printf("No segments left in SRH, would not forward.\n");
-          } else {
-            // Compute next-hop index such that ingress picks first segment and transit picks second
-            int next_sid_index = srh->last_entry - srh->segments_left + 1;
-            char next_sid_str[INET6_ADDRSTRLEN];
-            if (inet_ntop(AF_INET6, &srh->segments[next_sid_index], next_sid_str, sizeof(next_sid_str)) ==
-                NULL) {
-              perror("inet_ntop failed for next segment");
-            } else {
-              printf("Next segment IPv6 (from SRH): %s\n", next_sid_str);
-            }
-
-            // Lookup MAC for next segment
-            struct rte_ether_addr *next_mac = lookup_mac_for_ipv6(&srh->segments[next_sid_index]);
-            if (next_mac) {
-              printf("Resolved next-hop MAC: %02X:%02X:%02X:%02X:%02X:%02X\n", next_mac->addr_bytes[0],
-                     next_mac->addr_bytes[1], next_mac->addr_bytes[2], next_mac->addr_bytes[3],
-                     next_mac->addr_bytes[4], next_mac->addr_bytes[5]);
-            } else {
-              printf("No MAC mapping found for next segment IPv6!\n");
-            }
-
-            // Store original DstIP (ingress's own IP for this interface) to use as SrcIP later
-            struct in6_addr ingress_own_ip_as_src;
-            memcpy(&ingress_own_ip_as_src, &ipv6_hdr->dst_addr, sizeof(struct in6_addr));
-
-            // Update the main IPv6 header's Destination Address to the next segment.
-            memcpy(&ipv6_hdr->dst_addr, &srh->segments[next_sid_index], sizeof(struct in6_addr));
-
-            // Update the main IPv6 header's Source Address to the ingress node's own IP.
-            // memcpy(&ipv6_hdr->src_addr, &ingress_own_ip_as_src, sizeof(struct in6_addr));
-
-            // Get the MAC address of the ingress DPDK port (which will be the source MAC)
-            struct rte_ether_addr ingress_port_mac;
-            rte_eth_macaddr_get(rx_port_id, &ingress_port_mac);
-
-            // Print what the packet's destination/source MAC and IPv6 would be
-            printf("Packet would be sent with:\n");
-            printf("  Src MAC: %02X:%02X:%02X:%02X:%02X:%02X\n", ingress_port_mac.addr_bytes[0],
-                   ingress_port_mac.addr_bytes[1], ingress_port_mac.addr_bytes[2],
-                   ingress_port_mac.addr_bytes[3], ingress_port_mac.addr_bytes[4],
-                   ingress_port_mac.addr_bytes[5]);
-            if (next_mac) {
-              printf("  Dst MAC: %02X:%02X:%02X:%02X:%02X:%02X\n", next_mac->addr_bytes[0],
-                     next_mac->addr_bytes[1], next_mac->addr_bytes[2], next_mac->addr_bytes[3],
-                     next_mac->addr_bytes[4], next_mac->addr_bytes[5]);
-            }
-            printf("  Src IPv6: ");
-            print_ipv6_address((struct in6_addr *)&ipv6_hdr->src_addr, "");
-            printf("  Dst IPv6: ");
-            print_ipv6_address((struct in6_addr *)&ipv6_hdr->dst_addr, "");
-
-            // Actually send the packet
-            if (next_mac) {
-              send_packet_to(*next_mac, mbuf, rx_port_id);
-              printf("[INGRESS] Packet hex dump AFTER send (first 64 bytes):\n");
-              dump_len = rte_pktmbuf_pkt_len(mbuf);
-              if (dump_len > 64) dump_len = 64;
-              hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len);
-              printf("[INGRESS] After send: Packet length: %u\n", rte_pktmbuf_pkt_len(mbuf));
-              printf("[INGRESS] After send: mbuf nb_segs: %u\n", mbuf->nb_segs);
-              printf("[INGRESS] After send: tailroom: %u, data_len: %u\n", rte_pktmbuf_tailroom(mbuf),
-                     mbuf->data_len);
-            } else {
-              printf("No MAC address found for next segment, dropping packet\n");
-              rte_pktmbuf_free(mbuf);
-            }
-          }
-
-          break;
-        case 1:
-          // Bypass all operations
-          break;
-        case 2: add_custom_header_only(mbuf); break;
-        default: break;
-      }
-      break;
-    default: break;
-  }
-}
-
-// Modify process_ingress function to first check if packet is return traffic
-static inline void process_ingress(struct rte_mbuf **pkts, uint16_t nb_rx, uint16_t rx_port_id) {
-  printf("Starting process on ingress role with %u packets\n", nb_rx);
-  for (uint16_t i = 0; i < nb_rx; i++) {
-    struct rte_ether_hdr *eth_hdr = rte_pktmbuf_mtod(pkts[i], struct rte_ether_hdr *);
-    uint16_t ether_type = rte_be_to_cpu_16(eth_hdr->ether_type);
-
-    if (ether_type == RTE_ETHER_TYPE_IPV6) {
-      struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-
-      // Check if from iperf server (return traffic)
-      struct in6_addr iperf_server_addr;
-      inet_pton(AF_INET6, "2a05:d014:dc7:12c2:724:c0e1:c16d:2f16", &iperf_server_addr);
-      
-      char src_ip[INET6_ADDRSTRLEN];
-      inet_ntop(AF_INET6, &ipv6_hdr->src_addr, src_ip, INET6_ADDRSTRLEN);
-      printf("[INGRESS-CHECK] Checking if packet from %s is return traffic\n", src_ip);
-
-      if (memcmp(&ipv6_hdr->src_addr, &iperf_server_addr, sizeof(struct in6_addr)) == 0) {
-        // This is return traffic
-        printf("[INGRESS-CHECK] MATCH! Detected return traffic from iperf server\n");
-        process_ingress_return_packet(pkts[i]);
-      } else {
-        // This is client->server traffic
-        printf("[INGRESS-CHECK] Not a match, processing as normal client->server traffic\n");
-        process_ingress_packet(pkts[i], rx_port_id);
-      }
-    } else {
-      process_ingress_packet(pkts[i], rx_port_id);
-    }
-  }
-}
-// Add this function to handle return traffic at the transit node
-static inline void process_transit_return_packet(struct rte_mbuf *mbuf) {
-  struct rte_ether_hdr *eth_hdr = rte_pktmbuf_mtod(mbuf, struct rte_ether_hdr *);
-  struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-
-  printf("**** [TRANSIT-RETURN] DEBUG: Function called with packet length %u ****\n", rte_pktmbuf_pkt_len(mbuf));
-  printf("[TRANSIT-RETURN] MAC src: %02X:%02X:%02X:%02X:%02X:%02X\n", 
-         eth_hdr->src_addr.addr_bytes[0], eth_hdr->src_addr.addr_bytes[1],
-         eth_hdr->src_addr.addr_bytes[2], eth_hdr->src_addr.addr_bytes[3],
-         eth_hdr->src_addr.addr_bytes[4], eth_hdr->src_addr.addr_bytes[5]);
-  printf("[TRANSIT-RETURN] MAC dst: %02X:%02X:%02X:%02X:%02X:%02X\n",
-         eth_hdr->dst_addr.addr_bytes[0], eth_hdr->dst_addr.addr_bytes[1],
-         eth_hdr->dst_addr.addr_bytes[2], eth_hdr->dst_addr.addr_bytes[3],
-         eth_hdr->dst_addr.addr_bytes[4], eth_hdr->dst_addr.addr_bytes[5]);
-  
-  char src_ip[INET6_ADDRSTRLEN], dst_ip[INET6_ADDRSTRLEN];
-  inet_ntop(AF_INET6, &ipv6_hdr->src_addr, src_ip, INET6_ADDRSTRLEN);
-  inet_ntop(AF_INET6, &ipv6_hdr->dst_addr, dst_ip, INET6_ADDRSTRLEN);
-  
-  printf("[TRANSIT-RETURN] IPv6 src: %s\n", src_ip);
-  printf("[TRANSIT-RETURN] IPv6 dst: %s\n", dst_ip);
-  printf("[TRANSIT-RETURN] Next Header: %u\n", ipv6_hdr->proto);
-  
-  // Dump the first 64 bytes of the packet
-  printf("[TRANSIT-RETURN] Packet hex dump (first 64 bytes):\n");
-  size_t dump_len = rte_pktmbuf_pkt_len(mbuf);
-  if (dump_len > 64) dump_len = 64;
-  hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len);
-
-  // Forward to ingress node
-  struct rte_ether_addr ingress_mac = {
-      {0x02, 0x0c, 0xb4, 0x7a, 0x8c, 0x6d}};  // Replace with actual ingress MAC address
-
-  send_packet_to(ingress_mac, mbuf, 0);
-  printf("[TRANSIT-RETURN] Forwarded server response to ingress node\n");
-}
-
-// Helper: process a single packet for transit
-static inline void process_transit_packet(struct rte_mbuf *mbuf, int i) {
-  printf("[TRANSIT] Packet hex dump BEFORE processing (first 64 bytes):\n");
-  size_t dump_len = rte_pktmbuf_pkt_len(mbuf);
-  if (dump_len > 64) dump_len = 64;
-  hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len);
-  printf("[TRANSIT] Packet length: %u\n", rte_pktmbuf_pkt_len(mbuf));
-  printf("[TRANSIT] mbuf nb_segs: %u\n", mbuf->nb_segs);
-  printf("[TRANSIT] tailroom: %u, data_len: %u\n", rte_pktmbuf_tailroom(mbuf), mbuf->data_len);
-
-  struct rte_ether_hdr *eth_hdr = rte_pktmbuf_mtod(mbuf, struct rte_ether_hdr *);
-  uint16_t ether_type = rte_be_to_cpu_16(eth_hdr->ether_type);
-
-  // Skip non-unicast (multicast/broadcast) destination MAC
-  if ((eth_hdr->dst_addr.addr_bytes[0] & 0x01) != 0) {
-    printf("Transit: Multicast/Broadcast MAC, skipping and freeing mbuf\n");
-    // Print more info for debugging
-    printf("  MAC: %02X:%02X:%02X:%02X:%02X:%02X\n", eth_hdr->dst_addr.addr_bytes[0],
-           eth_hdr->dst_addr.addr_bytes[1], eth_hdr->dst_addr.addr_bytes[2], eth_hdr->dst_addr.addr_bytes[3],
-           eth_hdr->dst_addr.addr_bytes[4], eth_hdr->dst_addr.addr_bytes[5]);
-    struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-    printf("  IPv6 src: ");
-    print_ipv6_address((struct in6_addr *)&ipv6_hdr->src_addr, "");
-    printf("  IPv6 dst: ");
-    print_ipv6_address((struct in6_addr *)&ipv6_hdr->dst_addr, "");
-    printf("  EtherType: 0x%04x\n", ether_type);
-    printf("  Next header: %u\n", ipv6_hdr->proto);
-    if (ipv6_hdr->proto == IPPROTO_UDP) {
-      struct rte_udp_hdr *udp_hdr = (struct rte_udp_hdr *)(ipv6_hdr + 1);
-      printf("  UDP src port: %u, dst port: %u\n", ntohs(udp_hdr->src_port), ntohs(udp_hdr->dst_port));
-    } else if (ipv6_hdr->proto == IPPROTO_TCP) {
-      struct rte_tcp_hdr *tcp_hdr = (struct rte_tcp_hdr *)(ipv6_hdr + 1);
-      printf("  TCP src port: %u, dst port: %u\n", ntohs(tcp_hdr->src_port), ntohs(tcp_hdr->dst_port));
-    }
-    rte_pktmbuf_free(mbuf);
-    return;
-  }
-
-  switch (ether_type) {
-    case RTE_ETHER_TYPE_IPV6:
-      switch (operation_bypass_bit) {
-        case 0: {
-          struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-          struct ipv6_srh *srh = (struct ipv6_srh *)(ipv6_hdr + 1);
-
-          // Only process packets with your SRH
-          if (srh->next_header != 61 || srh->routing_type != 4) {
-            printf("Packet %d: Not a valid SRH packet, skipping\n", i + 1);
-            rte_pktmbuf_free(mbuf);
-            return;
-          }
-
-          if (srh->next_header == 61) {
-            printf("SRH detected at transit node \n");
-
-            // struct hmac_tlv *hmac = (struct hmac_tlv *)(srh + 1);
-            // struct pot_tlv *pot = (struct pot_tlv *)(hmac + 1);
-            // compute SRH total size in bytes
-            size_t srh_bytes = sizeof(struct ipv6_srh);
-            // hmac TLV follows immediately
-            uint8_t *hmac_ptr = (uint8_t *)srh + srh_bytes;
-            struct hmac_tlv *hmac = (struct hmac_tlv *)hmac_ptr;
-            // pot TLV follows right after the HMAC TLV
-            uint8_t *pot_ptr = hmac_ptr + sizeof(struct hmac_tlv);
-            struct pot_tlv *pot = (struct pot_tlv *)pot_ptr;
-
-            // Print offsets and hex dumps for debugging
-            print_offset_and_hex("[TRANSIT] SRH", eth_hdr, srh, sizeof(struct ipv6_srh));
-            print_offset_and_hex("[TRANSIT] HMAC TLV", eth_hdr, hmac, sizeof(struct hmac_tlv));
-            print_offset_and_hex("[TRANSIT] POT TLV", eth_hdr, pot, sizeof(struct pot_tlv));
-
-            printf("[TRANSIT] Raw HMAC TLV bytes: ");
-            hex_dump((void *)hmac, sizeof(struct hmac_tlv));
-            printf("[TRANSIT] hmac->hmac_key_id: %u\n", rte_be_to_cpu_32(hmac->hmac_key_id));
-
-            // Display source and destination MAC addresses
-            printf("Packet %d:\n", i + 1);
-            printf("  Src MAC: %02" PRIx8 ":%02" PRIx8 ":%02" PRIx8 ":%02" PRIx8 ":%02" PRIx8 ":%02" PRIx8
-                   "\n",
-                   eth_hdr->src_addr.addr_bytes[0], eth_hdr->src_addr.addr_bytes[1],
-                   eth_hdr->src_addr.addr_bytes[2], eth_hdr->src_addr.addr_bytes[3],
-                   eth_hdr->src_addr.addr_bytes[4], eth_hdr->src_addr.addr_bytes[5]);
-            printf("  Dst MAC: %02" PRIx8 ":%02" PRIx8 ":%02" PRIx8 ":%02" PRIx8 ":%02" PRIx8 ":%02" PRIx8
-                   "\n",
-                   eth_hdr->dst_addr.addr_bytes[0], eth_hdr->dst_addr.addr_bytes[1],
-                   eth_hdr->dst_addr.addr_bytes[2], eth_hdr->dst_addr.addr_bytes[3],
-                   eth_hdr->dst_addr.addr_bytes[4], eth_hdr->dst_addr.addr_bytes[5]);
-            printf("  EtherType: 0x%04x\n", rte_be_to_cpu_16(eth_hdr->ether_type));
-
-            print_ipv6_address((struct in6_addr *)&ipv6_hdr->src_addr, "source");
-            print_ipv6_address((struct in6_addr *)&ipv6_hdr->dst_addr, "destination");
-
-            printf("The size of srh is %lu\n", sizeof(*srh));
-            printf("The size of hmac is %lu\n", sizeof(*hmac));
-            printf("The size of pot is %lu\n", sizeof(*pot));
-
-            printf("HMAC type: %u\n", hmac->type);
-            printf("HMAC length: %u\n", hmac->length);
-            printf("HMAC key ID: %u\n", rte_be_to_cpu_32(hmac->hmac_key_id));
-            printf("HMAC size: %ld\n", sizeof(hmac->hmac_value));
-
-            // Decrypt the POT field (PVF) if needed
-            char dst_ip_str[INET6_ADDRSTRLEN];
-            if (inet_ntop(AF_INET6, &ipv6_hdr->dst_addr, dst_ip_str, sizeof(dst_ip_str)) == NULL) {
-              perror("inet_ntop failed");
-              rte_pktmbuf_free(mbuf);
-              return;
-            }
-
-            // uint8_t k_pot_in_transit[1][HMAC_MAX_LENGTH] = {{0}};  // Changed declaration
-            // if (read_encryption_key("keys.txt", dst_ip_str, k_pot_in_transit[0], HMAC_MAX_LENGTH) !=
-            //     0) {  // Read into k_pot_in_transit[0]
-            //   printf("Failed to read key for %s\n", dst_ip_str);
-            //   rte_pktmbuf_free(mbuf);
-            //   return;
-            // }
-
-            printf("Decrypting PVF for %s\n", dst_ip_str);
-            uint8_t pvf_out[HMAC_MAX_LENGTH];
-            // memcpy(pvf_out, pot->encrypted_hmac, HMAC_MAX_LENGTH);
-            // decrypt_pvf(k_pot_in_transit, pot->nonce, pvf_out);  // Use k_pot_in_transit
-            // memcpy(pot->encrypted_hmac, pvf_out, HMAC_MAX_LENGTH);
-            memcpy(pvf_out, pot->encrypted_hmac, HMAC_MAX_LENGTH);
-            // decrypt_pvf(k_pot_in_transit, pot->nonce, pvf_out);
-            decrypt_pvf(&k_pot_in[1], pot->nonce, pvf_out);
-
-            // memcpy(hmac->hmac_value, pvf_out, HMAC_MAX_LENGTH);
-            memcpy(pot->encrypted_hmac, pvf_out, HMAC_MAX_LENGTH);
-            printf("Transit: Updated HMAC field with decrypted PVF\n");
-            printf("[TRANSIT] Decrypted PVF:\n");
-            for (int j = 0; j < HMAC_MAX_LENGTH; j++) printf("%02x", pvf_out[j]);
-            printf("\n[TRANSIT] HMAC field after update:\n");
-            for (int j = 0; j < HMAC_MAX_LENGTH; j++) printf("%02x", hmac->hmac_value[j]);
-            printf("\n");
-
-            // SRH forwarding logic
-            if (srh->segments_left == 0) {
-              printf("Transit: No more segments left, dropping or processing as egress\n");
-              rte_pktmbuf_free(mbuf);
-              return;
-            }
-
-            // Decrement segments_left before computing next hop
-            srh->segments_left--;
-            // Compute next-hop index: = last_entry - segments_left + 1
-            int next_sid_index = srh->last_entry - srh->segments_left + 1;
-            printf("Transit: Forwarding to next segment, segments left: %d, next index: %d\n",
-                   srh->segments_left, next_sid_index);
-            // Update IPv6 dst to selected segment
-            memcpy(&ipv6_hdr->dst_addr, &srh->segments[next_sid_index], sizeof(ipv6_hdr->dst_addr));
-            // Lookup MAC for new IPv6 destination
-            struct rte_ether_addr *next_mac = lookup_mac_for_ipv6(&srh->segments[next_sid_index]);
-            char next_ip_str[INET6_ADDRSTRLEN];
-            printf("Transit: Next segment IPv6: %s\n",
-                   inet_ntop(AF_INET6, &srh->segments[next_sid_index], next_ip_str, sizeof(next_ip_str)));
-            if (next_mac) {
-              printf("Transit: Found MAC for next segment: %02X:%02X:%02X:%02X:%02X:%02X\n",
-                     next_mac->addr_bytes[0], next_mac->addr_bytes[1], next_mac->addr_bytes[2],
-                     next_mac->addr_bytes[3], next_mac->addr_bytes[4], next_mac->addr_bytes[5]);
-              send_packet_to(*next_mac, mbuf, /*tx_port_id*/ 0);
-              printf("[TRANSIT] Packet hex dump AFTER send (first 64 bytes):\n");
-              dump_len = rte_pktmbuf_pkt_len(mbuf);
-              if (dump_len > 64) dump_len = 64;
-              hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len);
-              printf("[TRANSIT] After send: Packet length: %u\n", rte_pktmbuf_pkt_len(mbuf));
-              printf("[TRANSIT] After send: mbuf nb_segs: %u\n", mbuf->nb_segs);
-              printf("[TRANSIT] After send: tailroom: %u, data_len: %u\n", rte_pktmbuf_tailroom(mbuf),
-                     mbuf->data_len);
-            } else {
-              printf("Transit: No MAC mapping for next segment!\n");
-              rte_pktmbuf_free(mbuf);
-            }
-          }
-          break;
-        }  // Added braces for case block
-        case 1:
-          // Bypass all operations
-          break;
-        // case 2: remove_headers_only(mbuf); break;
-        default: break;
-      }
-      break;
-    default: break;
-  }
-}
-
-// Modify process_transit function to first check if packet is return traffic
-static inline void process_transit(struct rte_mbuf **pkts, uint16_t nb_rx) {
-  printf("**** [TRANSIT] Starting transit processing with %u packets ****\n", nb_rx);
-  for (uint16_t i = 0; i < nb_rx; i++) {
-    struct rte_ether_hdr *eth_hdr = rte_pktmbuf_mtod(pkts[i], struct rte_ether_hdr *);
-    uint16_t ether_type = rte_be_to_cpu_16(eth_hdr->ether_type);
-
-    printf("[TRANSIT] Processing packet %u, EtherType: 0x%04x\n", i, ether_type);
-    
-    if (ether_type == RTE_ETHER_TYPE_IPV6) {
-      struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-
-      // Check if from iperf server (return traffic)
-      struct in6_addr iperf_server_addr;
-      inet_pton(AF_INET6, "2a05:d014:dc7:12c2:724:c0e1:c16d:2f16", &iperf_server_addr);
-
-      char src_ip[INET6_ADDRSTRLEN];
-      inet_ntop(AF_INET6, &ipv6_hdr->src_addr, src_ip, INET6_ADDRSTRLEN);
-      printf("[TRANSIT-CHECK] Checking if packet from %s is return traffic\n", src_ip);
-
-      if (memcmp(&ipv6_hdr->src_addr, &iperf_server_addr, sizeof(struct in6_addr)) == 0) {
-        // This is return traffic
-        printf("[TRANSIT-CHECK] MATCH! Detected return traffic from iperf server\n");
-        process_transit_return_packet(pkts[i]);
-      } else {
-        // This is client->server traffic
-        printf("[TRANSIT-CHECK] Not a match, processing as normal client->server traffic\n");
-        process_transit_packet(pkts[i], i);
-      }
-    } else {
-      process_transit_packet(pkts[i], i);
-    }
-  }
-}
 
 int compare_hmac(struct hmac_tlv *hmac, uint8_t *hmac_out, struct rte_mbuf *mbuf) {
   if (memcmp(hmac->hmac_value, hmac_out, 32) != 0) {
@@ -887,269 +299,6 @@ int compare_hmac(struct hmac_tlv *hmac, uint8_t *hmac_out, struct rte_mbuf *mbuf
     printf("The transit of the packet is verified\n");
     // forward it to the tap interface so iperf can catch it
     return 1;
-  }
-}
-
-// Add this function to handle return traffic at the egress node
-static inline void process_return_traffic(struct rte_mbuf *mbuf) {
-  struct rte_ether_hdr *eth_hdr = rte_pktmbuf_mtod(mbuf, struct rte_ether_hdr *);
-  struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-
-  printf("**** [EGRESS-RETURN] DEBUG: Function called with packet length %u ****\n", rte_pktmbuf_pkt_len(mbuf));
-  printf("[EGRESS-RETURN] MAC src: %02X:%02X:%02X:%02X:%02X:%02X\n", 
-         eth_hdr->src_addr.addr_bytes[0], eth_hdr->src_addr.addr_bytes[1],
-         eth_hdr->src_addr.addr_bytes[2], eth_hdr->src_addr.addr_bytes[3],
-         eth_hdr->src_addr.addr_bytes[4], eth_hdr->src_addr.addr_bytes[5]);
-  printf("[EGRESS-RETURN] MAC dst: %02X:%02X:%02X:%02X:%02X:%02X\n",
-         eth_hdr->dst_addr.addr_bytes[0], eth_hdr->dst_addr.addr_bytes[1],
-         eth_hdr->dst_addr.addr_bytes[2], eth_hdr->dst_addr.addr_bytes[3],
-         eth_hdr->dst_addr.addr_bytes[4], eth_hdr->dst_addr.addr_bytes[5]);
-
-  // Get source and destination IP addresses
-  char src_ip[INET6_ADDRSTRLEN], dst_ip[INET6_ADDRSTRLEN];
-  inet_ntop(AF_INET6, &ipv6_hdr->src_addr, src_ip, INET6_ADDRSTRLEN);
-  inet_ntop(AF_INET6, &ipv6_hdr->dst_addr, dst_ip, INET6_ADDRSTRLEN);
-  
-  printf("[EGRESS-RETURN] IPv6 src: %s\n", src_ip);
-  printf("[EGRESS-RETURN] IPv6 dst: %s\n", dst_ip);
-  printf("[EGRESS-RETURN] Next Header: %u\n", ipv6_hdr->proto);
-
-  // Check if this is from the iperf server
-  struct in6_addr iperf_server_addr;
-  inet_pton(AF_INET6, "2a05:d014:dc7:12c2:724:c0e1:c16d:2f16", &iperf_server_addr);
-
-  // Print additional UPD/TCP header info if applicable
-  if (ipv6_hdr->proto == IPPROTO_UDP) {
-    struct rte_udp_hdr *udp_hdr = (struct rte_udp_hdr *)(ipv6_hdr + 1);
-    printf("[EGRESS-RETURN] UDP src port: %u, dst port: %u, dgram_len: %u\n", 
-           ntohs(udp_hdr->src_port), ntohs(udp_hdr->dst_port), ntohs(udp_hdr->dgram_len));
-  } else if (ipv6_hdr->proto == IPPROTO_TCP) {
-    struct rte_tcp_hdr *tcp_hdr = (struct rte_tcp_hdr *)(ipv6_hdr + 1);
-    printf("[EGRESS-RETURN] TCP src port: %u, dst port: %u\n",
-           ntohs(tcp_hdr->src_port), ntohs(tcp_hdr->dst_port));
-  } else if (ipv6_hdr->proto == IPPROTO_ICMPV6) {
-    printf("[EGRESS-RETURN] ICMPv6 packet detected\n");
-  }
-
-  if (memcmp(&ipv6_hdr->src_addr, &iperf_server_addr, sizeof(struct in6_addr)) == 0) {
-    printf("[EGRESS-RETURN] Detected iperf server response\n");
-
-    // Dump the first 64 bytes of the packet
-    printf("[EGRESS-RETURN] Packet hex dump (first 64 bytes):\n");
-    size_t dump_len = rte_pktmbuf_pkt_len(mbuf);
-    if (dump_len > 64) dump_len = 64;
-    hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len);
-
-    // Find transit node MAC (first hop in return path)
-    struct rte_ether_addr transit_mac = {
-        {0x02, 0xf5, 0x27, 0x51, 0xbc, 0x1d}};  // Replace with your transit MAC
-
-    // Forward packet to transit node
-    send_packet_to(transit_mac, mbuf, 0);
-    printf("[EGRESS-RETURN] Forwarded server response to transit node\n");
-  }
-}
-
-// Helper: process a single packet for egress
-static inline void process_egress_packet(struct rte_mbuf *mbuf) {
-  printf("[EGRESS] Packet hex dump BEFORE processing (first 64 bytes):\n");
-  size_t dump_len = rte_pktmbuf_pkt_len(mbuf);
-  if (dump_len > 64) dump_len = 64;
-  hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len);
-  printf("[EGRESS] Packet length: %u\n", rte_pktmbuf_pkt_len(mbuf));
-  printf("[EGRESS] mbuf nb_segs: %u\n", mbuf->nb_segs);
-  printf("[EGRESS] tailroom: %u, data_len: %u\n", rte_pktmbuf_tailroom(mbuf), mbuf->data_len);
-
-  struct rte_ether_hdr *eth_hdr = rte_pktmbuf_mtod(mbuf, struct rte_ether_hdr *);
-  printf("EGRESS DEBUG: First 32 bytes: ");
-  hex_dump(rte_pktmbuf_mtod(mbuf, void *), 32);
-  printf("EGRESS DEBUG: Packet length: %u\n", rte_pktmbuf_pkt_len(mbuf));
-  printf("EGRESS DEBUG: EtherType: 0x%04x\n", rte_be_to_cpu_16(eth_hdr->ether_type));
-  uint16_t ether_type = rte_be_to_cpu_16(eth_hdr->ether_type);
-
-  switch (ether_type) {
-    case RTE_ETHER_TYPE_IPV6:
-      switch (operation_bypass_bit) {
-        case 0: {
-          struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-          struct ipv6_srh *srh = (struct ipv6_srh *)(ipv6_hdr + 1);
-
-          if (srh->next_header == 61) {
-            printf("Egress: SRH detected, processing packet\n");
-
-            size_t srh_bytes = sizeof(struct ipv6_srh);
-            uint8_t *hmac_ptr = (uint8_t *)srh + srh_bytes;
-            struct hmac_tlv *hmac = (struct hmac_tlv *)hmac_ptr;
-            uint8_t *pot_ptr = hmac_ptr + sizeof(struct hmac_tlv);
-            struct pot_tlv *pot = (struct pot_tlv *)pot_ptr;
-
-            // Print offsets and hex dumps for debugging
-            print_offset_and_hex("[EGRESS] SRH", eth_hdr, srh, sizeof(struct ipv6_srh));
-            print_offset_and_hex("[EGRESS] HMAC TLV", eth_hdr, hmac, sizeof(struct hmac_tlv));
-            print_offset_and_hex("[EGRESS] POT TLV", eth_hdr, pot, sizeof(struct pot_tlv));
-
-            // Read key for this egress node from keys.txt
-            char dst_ip_str[INET6_ADDRSTRLEN];
-            if (inet_ntop(AF_INET6, &ipv6_hdr->dst_addr, dst_ip_str, sizeof(dst_ip_str)) == NULL) {
-              perror("inet_ntop failed");
-              rte_pktmbuf_free(mbuf);
-              return;
-            }
-
-            // uint8_t k_pot_in_egress[1][HMAC_MAX_LENGTH] = {{0}};
-            // if (read_encryption_key("keys.txt", dst_ip_str, k_pot_in_egress[0], HMAC_MAX_LENGTH) != 0) {
-            //   printf("Egress: Failed to read key for %s\n", dst_ip_str);
-            //   rte_pktmbuf_free(mbuf);
-            //   return;
-            // }
-
-            // Decrypt PVF
-            uint8_t hmac_out[HMAC_MAX_LENGTH];
-            memcpy(hmac_out, pot->encrypted_hmac, HMAC_MAX_LENGTH);
-            // decrypt_pvf(k_pot_in_egress, pot->nonce, hmac_out);
-            decrypt_pvf(&k_pot_in[0], pot->nonce, hmac_out);
-            memcpy(pot->encrypted_hmac, hmac_out, HMAC_MAX_LENGTH);
-            printf("[EGRESS] Decrypted PVF:\n");
-            for (int j = 0; j < HMAC_MAX_LENGTH; j++) printf("%02x", hmac_out[j]);
-            printf("\n[EGRESS] HMAC field in header:\n");
-            for (int j = 0; j < HMAC_MAX_LENGTH; j++) printf("%02x", hmac->hmac_value[j]);
-            printf("\n");
-
-            // Prepare HMAC key (use same as encryption key for demo)
-            // uint8_t k_hmac_ie[HMAC_MAX_LENGTH] = {0};
-            // if (read_encryption_key("keys.txt", dst_ip_str, k_hmac_ie, HMAC_MAX_LENGTH) != 0) {
-            //   printf("Egress: Failed to read HMAC key for %s\n", dst_ip_str);
-            //   rte_pktmbuf_free(mbuf);
-            //   return;
-            // }
-            uint8_t *k_hmac_ie = k_pot_in[0];  // egress key
-
-            // Log HMAC input values for debugging
-            printf("[EGRESS] HMAC input values:\n");
-            printf("  src_addr: ");
-            char src_addr_str[INET6_ADDRSTRLEN];
-            if (inet_ntop(AF_INET6, &ipv6_hdr->src_addr, src_addr_str, sizeof(src_addr_str)))
-              printf("%s\n", src_addr_str);
-            else
-              perror("inet_ntop src_addr");
-            printf("  srh->last_entry: %u\n", srh->last_entry);
-            printf("  srh->flags: %u\n", srh->flags);
-            printf("  hmac->hmac_key_id: %u\n", rte_be_to_cpu_32(hmac->hmac_key_id));
-            printf("  srh->segments: ");
-            for (int i = 0; i <= srh->last_entry; i++) {
-              char seg_str[INET6_ADDRSTRLEN];
-              if (inet_ntop(AF_INET6, &srh->segments[i], seg_str, sizeof(seg_str)))
-                printf("%s ", seg_str);
-              else
-                perror("inet_ntop segment");
-            }
-            printf("\n");
-            uint8_t expected_hmac[HMAC_MAX_LENGTH];
-            if (calculate_hmac((uint8_t *)&ipv6_hdr->src_addr, srh, hmac, k_hmac_ie, HMAC_MAX_LENGTH,
-                               expected_hmac) != 0) {
-              printf("[EGRESS] Expected HMAC:\n");
-              for (int j = 0; j < HMAC_MAX_LENGTH; j++) printf("%02x", expected_hmac[j]);
-              printf("\n");
-              printf("Egress: HMAC calculation failed\n");
-              rte_pktmbuf_free(mbuf);
-              return;
-            }
-
-            // Compare decrypted PVF with expected HMAC
-            if (memcmp(hmac_out, expected_hmac, HMAC_MAX_LENGTH) != 0) {
-              printf("Egress: HMAC verification failed, dropping packet\n");
-              rte_pktmbuf_free(mbuf);
-              return;
-            } else {
-              printf("Egress: HMAC verified successfully, forwarding packet\n");
-            }
-
-            // Remove headers and forward to iperf server (replace MAC/port as needed)
-            remove_headers(mbuf);
-            
-            // Debug iperf traffic specifically
-            struct rte_ether_hdr *eth_hdr_final = rte_pktmbuf_mtod(mbuf, struct rte_ether_hdr *);
-            struct rte_ipv6_hdr *ipv6_hdr_final = (struct rte_ipv6_hdr *)(eth_hdr_final + 1);
-            
-            // Get source and destination IP addresses after header removal
-            char final_src_ip[INET6_ADDRSTRLEN], final_dst_ip[INET6_ADDRSTRLEN];
-            inet_ntop(AF_INET6, &ipv6_hdr_final->src_addr, final_src_ip, INET6_ADDRSTRLEN);
-            inet_ntop(AF_INET6, &ipv6_hdr_final->dst_addr, final_dst_ip, INET6_ADDRSTRLEN);
-            
-            printf("[EGRESS-IPERF-DEBUG] Final packet IPv6 src: %s\n", final_src_ip);
-            printf("[EGRESS-IPERF-DEBUG] Final packet IPv6 dst: %s\n", final_dst_ip);
-            
-            // Check for UDP traffic to iperf port
-            if (ipv6_hdr_final->proto == IPPROTO_UDP) {
-                struct rte_udp_hdr *udp_hdr = (struct rte_udp_hdr *)(ipv6_hdr_final + 1);
-                printf("[EGRESS-IPERF-DEBUG] UDP packet: src_port=%u, dst_port=%u, dgram_len=%u\n", 
-                       ntohs(udp_hdr->src_port), ntohs(udp_hdr->dst_port), ntohs(udp_hdr->dgram_len));
-                
-                if (ntohs(udp_hdr->dst_port) == 5201) {
-                    printf("[EGRESS-IPERF-DEBUG] ****** DETECTED UDP PACKET TO IPERF PORT 5201 ******\n");
-                    printf("[EGRESS-IPERF-DEBUG] Watch for server response in tcpdump\n");
-                }
-            }
-            
-            // struct rte_ether_addr iperf_mac = {{0x08, 0x00, 0x27, 0x7D, 0xDD, 0x01}};
-            struct rte_ether_addr iperf_mac = {
-                {0x02, 0x38, 0x81, 0xe2, 0xf9, 0xa7}};  // Updated to your iperf server MAC
-            // send_packet_to(iperf_mac, mbuf, /*tx_port_id*/ 1);
-            printf("[EGRESS-IPERF-DEBUG] Sending to iperf server MAC: %02X:%02X:%02X:%02X:%02X:%02X\n", 
-                   iperf_mac.addr_bytes[0], iperf_mac.addr_bytes[1], iperf_mac.addr_bytes[2], 
-                   iperf_mac.addr_bytes[3], iperf_mac.addr_bytes[4], iperf_mac.addr_bytes[5]);
-            send_packet_to(iperf_mac, mbuf, 0);
-            printf("[EGRESS] Packet hex dump AFTER send (first 64 bytes):\n");
-            dump_len = rte_pktmbuf_pkt_len(mbuf);
-            if (dump_len > 64) dump_len = 64;
-            hex_dump(rte_pktmbuf_mtod(mbuf, void *), dump_len);
-            printf("[EGRESS] After send: Packet length: %u\n", rte_pktmbuf_pkt_len(mbuf));
-            printf("[EGRESS] After send: mbuf nb_segs: %u\n", mbuf->nb_segs);
-            printf("[EGRESS] After send: tailroom: %u, data_len: %u\n", rte_pktmbuf_tailroom(mbuf),
-                   mbuf->data_len);
-          }
-          break;
-        }
-        case 1:
-          // Bypass all operations
-          break;
-        // case 2: remove_headers_only(mbuf); break;
-        default: break;
-      }
-      break;
-    default: break;
-  }
-}
-
-// Modify process_egress function to first check if packet is return traffic
-static inline void process_egress(struct rte_mbuf **pkts, uint16_t nb_rx) {
-  for (uint16_t i = 0; i < nb_rx; i++) {
-    struct rte_ether_hdr *eth_hdr = rte_pktmbuf_mtod(pkts[i], struct rte_ether_hdr *);
-    uint16_t ether_type = rte_be_to_cpu_16(eth_hdr->ether_type);
-
-    if (ether_type == RTE_ETHER_TYPE_IPV6) {
-      struct rte_ipv6_hdr *ipv6_hdr = (struct rte_ipv6_hdr *)(eth_hdr + 1);
-
-      // Check if packet is from iperf server
-      struct in6_addr iperf_server_addr;
-      inet_pton(AF_INET6, "2a05:d014:dc7:12c2:724:c0e1:c16d:2f16", &iperf_server_addr);
-      
-      char src_ip[INET6_ADDRSTRLEN];
-      inet_ntop(AF_INET6, &ipv6_hdr->src_addr, src_ip, INET6_ADDRSTRLEN);
-      printf("[EGRESS-CHECK] Checking if packet from %s is return traffic\n", src_ip);
-
-      if (memcmp(&ipv6_hdr->src_addr, &iperf_server_addr, sizeof(struct in6_addr)) == 0) {
-        // This is return traffic
-        printf("[EGRESS-CHECK] MATCH! Detected return traffic from iperf server\n");
-        process_return_traffic(pkts[i]);
-      } else {
-        // This is client->server traffic
-        printf("[EGRESS-CHECK] Not a match, processing as normal client->server traffic\n");
-        process_egress_packet(pkts[i]);
-      }
-    } else {
-      process_egress_packet(pkts[i]);
-    }
   }
 }
